Add /thanks redirect
Two disadvantage of sending the HTML back directly in the POST are that:
a) We can't send someone the address of it (it's the same URL as the
submission form), and,
b) We can't refresh the page without resubmitting the form.

A redirect to show the success of submitting the form allows us to do
both a) and b).