Split out Models & Handlers into their own files
A migration towards a neater code organisation

Step 1: A file for each DB Model and a file for each Handler (now)
Step 2: Modules for these -- 1 module for all the Models and 1 module
for all the Handlers
Step 3: Will involve readdressing the inheritance hierarchy of the
Handlers. The previous steps should make the hierarchy more apparent and
easier to implement.